+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Auto suggest filename based on texture name
+- Auto update mutator name based on type
 
 ## [1.1.0] - 2021-04-27